move fairRand to code.py, incorporate right random lib
